% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_to_traits.R
\name{annot_to_traits}
\alias{annot_to_traits}
\title{Converts a tabular annotation file to a trait table ready to be used for imputation}
\usage{
annot_to_traits(annot, genome_index, kegg_index)
}
\arguments{
\item{annot}{A tabular functional genome annotation file.}

\item{genome_index}{Column number in which the genome name is indicated}

\item{kegg_index}{Column number in which to search for KEGG identifiers}
}
\value{
A trait presence/absence table
}
\description{
Converts a tabular annotation file to a trait table ready to be used for imputation
}
\examples{
imputation(imputation)
}
