% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completer.R
\name{completer}
\alias{completer}
\title{Imputes gene annotations to incomplete genomes based on phylogenetic signal.}
\usage{
completer(traits, tree, focal_genomes, span = 100, power = 2, threshold = 0.9)
}
\arguments{
\item{traits}{Table containing presence/absence information of gene annotations (columns) across genomes (rows).}

\item{tree}{GTDB tree containing target and reference genomes.}

\item{focal_genomes}{Optional vector of focal genomes to conduct imputation. If not provided, target genomes are inferred from the tree.}

\item{span}{Number of neighboring genomes to be employed for calculating the imputed values.}

\item{power}{Power value used when assigning weights to neighbor genomes (larger values assign larger weights to close neighbors)}

\item{threshold}{Relative threshold to assign a presence.}
}
\value{
A table of imputed values and imputation statistics.
}
\description{
Imputes gene annotations to incomplete genomes based on phylogenetic signal.
}
\examples{
completer(genome_traits,tree)
}
