% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfocaltree.R
\name{getfocaltree}
\alias{getfocaltree}
\title{Extracts a focal tree containing a certain number of closest neighbor tips from a larger tree}
\usage{
getfocaltree(focal_tip, tip_distances, tree, span = 100, power = 2)
}
\arguments{
\item{focal_tip}{Name of the focal tip around which the tree will be subseted.}

\item{tip_distances}{Distance matrix indicating the pairwaise tip distances in the reference tree.}

\item{tree}{The reference tree}

\item{span}{Number of neighboring genomes to be employed for calculating the imputed values.}

\item{power}{Power value used when assigning weights to neighbor genomes (larger values assign larger weights to close neighbors)}
}
\value{
A subseted tree
}
\description{
Extracts a focal tree containing a certain number of closest neighbor tips from a larger tree
}
\examples{
getfocaltree("MAG014122",tip_distances,tree)
}
