% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impsens.R
\name{impsens}
\alias{impsens}
\title{Calculates sensitivity of the imputation}
\usage{
impsens(imputation)
}
\arguments{
\item{imputation}{An imputation object generated by imputer}
}
\value{
A sensitivity value
}
\description{
Calculates sensitivity of the imputation
}
\examples{
imputation(imputation)
}
