% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{imputer}
\alias{imputer}
\title{Imputes gene annotations to incomplete genomes based on phylogenetic signal.}
\usage{
imputer(
  traits,
  focaltree,
  focal_genome,
  span = span,
  power = power,
  threshold = threshold
)
}
\arguments{
\item{traits}{Table containing presence/absence information of gene annotations (columns) across genomes (rows).}

\item{focaltree}{The subseted tree generated by getfocaltree based on the focal genome}

\item{focal_genome}{The name of the focal genome to be imputed.}

\item{span}{Number of neighboring genomes to be employed for calculating the imputed values.}

\item{power}{Power value used when assigning weights to neighbor genomes (larger values assign larger weights to close neighbors)}

\item{threshold}{Relative threshold to assign a presence.}
}
\value{
A table of imputed values and imputation statistics.
}
\description{
Imputes gene annotations to incomplete genomes based on phylogenetic signal.
}
\examples{
completer(genome_traits,tree)
}
