% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incompleter.R
\name{incompleter}
\alias{incompleter}
\title{Filters out annotation table to emulate lower completeness genomes}
\usage{
incompleter(
  annotation,
  genome_index,
  contig_index,
  gene_index,
  completeness,
  iter = 1
)
}
\arguments{
\item{genome_index}{Column number in which the genome name is indicated}

\item{contig_index}{Column number in which the contig name is indicated}

\item{gene_index}{Column number in which the gene name is indicated}

\item{completeness}{Completeness percentage to be emulated}

\item{iter}{Number of iterations. In each iteration a random set of contigs will be removed}

\item{annot}{A tabular functional genome annotation file.}
}
\value{
A table of imputed values and imputation statistics.
}
\description{
Filters out annotation table to emulate lower completeness genomes
}
\examples{
completer(genome_traits,tree)
}
