% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impsens.R
\name{recovery}
\alias{recovery}
\title{Calculates partial recovery of the imputation}
\usage{
recovery(imputation)
}
\arguments{
\item{imputation}{An imputation object generated by imputer}
}
\value{
A sensitivity value
}
\description{
Calculates partial recovery of the imputation
}
\examples{
recovery(imputation)
}
