% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_tree.R
\name{references_tree}
\alias{references_tree}
\title{Generate a vector of closest reference genomes}
\usage{
references_tree(focal_genomes, user_tree, reference_tree)
}
\arguments{
\item{focal_genomes}{Vector of focal genomes}

\item{user_tree}{User reference tree}

\item{reference_tree}{CompleteR reference tree}
}
\value{
A vector of closest reference genomes
}
\description{
Generate a vector of closest reference genomes
}
