% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_completer.R
\name{setup_completer}
\alias{setup_completer}
\title{Prepare reference data for CompleteR}
\usage{
setup_completer()
}
\description{
Prepare reference data for CompleteR
}
\examples{
imputation(imputation)
}
